refactor: move inner proc macro functions to shared
